requirements/base.txt
# Core dependencies
tkinter>=8.6
keyring>=23.0.0
psutil>=5.9.0
PyGithub>=1.55
python-crontab>=2.6.0
cryptography>=41.0.0
requests>=2.31.0
python-dateutil>=2.8.2 